divide modules between state and event repositories
