+++ src/decider/repository/in_memory/versioned_with_streams/error.rs
@@ -1,0 +1,7 @@
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+pub enum Error {
+    #[error("Cannot append, event version is out of date")]
+    VersionConflict,
+}
